Fix warning in register-table abstraction when newlib is libc
Building ufw uses -Wall and -Werror on compilers that support it. Building
cross-compiling to arm-none-eabi, it is common to use newlib as the toolchain's
standard C library.

Here is how a CI instance barfed when it tried to build fafc10a70:

  [ 29%] Building C object deps/ufw/CMakeFiles/ufw.dir/src/c/register-table.c.obj
  deps/ufw/src/c/register-table.c: In function 'reg_is_hexstr':
  deps/ufw/src/c/register-table.c:686:9: error: array subscript has type 'char' [-Werror=char-subscripts]
           if (isxdigit(s[idx]) == false)
           ^
  deps/ufw/src/c/register-table.c: In function 'reg_atom_from_hexstr':
  deps/ufw/src/c/register-table.c:723:9: error: array subscript has type 'char' [-Werror=char-subscripts]
           const RegisterAtom v = reg_c2a(tolower(s[idx]));
           ^
  cc1: all warnings being treated as errors

…needless to say, I was surprised. Here is why this happens:

  #ifndef __cplusplus
  /* These macros are intentionally written in a manner that will trigger
     a gcc -Wall warning if the user mistakenly passes a 'char' instead
     of an int containing an 'unsigned char'.  Note that the sizeof will
     always be 1, which is what we want for mapping EOF to __ctype_ptr__[0];
     the use of a raw index inside the sizeof triggers the gcc warning if
     __c was of type char, and sizeof masks side effects of the extra __c.
     Meanwhile, the real index to __ctype_ptr__+1 must be cast to int,
     since isalpha(0x100000001LL) must equal isalpha(1), rather than being
     an out-of-bounds reference on a 64-bit machine.  */
  #define __ctype_lookup(__c) ((__ctype_ptr__+sizeof(""[__c]))[(int)(__c)])
  [...]
  #define isxdigit(__c)   (__ctype_lookup(__c)&(_X|_N))
  [...]
  #endif

That __ctype_lookup() macro is one of the uglier ones I've seen. But it does
what is advertised. The tolower() case is similar. I'm not quite sure these are
legitimate concerns and having to write isxdigit((int)c) is ugly as all hell.
But I guess there is not a lot we can do about that.

+++ src/c/register-table.c
@@ -683,7 +683,7 @@ static bool
 reg_is_hexstr(const char *s, const size_t n)
 {
     for (size_t idx = 0u; idx < n; ++idx) {
-        if (isxdigit(s[idx]) == false)
+        if (isxdigit((int)s[idx]) == false)
             return false;
     }
 
@@ -720,7 +720,7 @@ reg_atom_from_hexstr(const char *s, const size_t n)
 
     for (size_t idx = 0u; idx < n; ++idx) {
         const size_t shift = (n-idx-1) * 4u;
-        const RegisterAtom v = reg_c2a(tolower(s[idx]));
+        const RegisterAtom v = reg_c2a(tolower((int)s[idx]));
         rv |= (v & 0x0fu) << shift;
     }
